# Improve the performance of hash Spark functions for scalar types through auto-vectorization