Improve the performance of hash Spark functions for scalar types through auto-vectorization.
We can observe a measurable speed-up with this change: 
TINYINT: ~4X, SMALLINT: ~4X, BIGINT: ~2X, REAL: ~4X, DOUBLE:~2X, DECIMAL(18, 6): ~2X, DECIMAL(38, 6): ~1.2X, TIMESTAMP: ~1.8X, VARCHAR: ~1.3X.
Benchmark result before this optimization:
```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
hash_ARRAY<MAP<INTEGER,VARCHAR>>##hash                      9.63ms    103.88
hash_ARRAY<MAP<INTEGER,VARCHAR>>##xxhash64                  9.38ms    106.66
----------------------------------------------------------------------------
hash_BIGINT##hash                                         352.84us     2.83K
hash_BIGINT##xxhash64                                     346.62us     2.88K
----------------------------------------------------------------------------
hash_DECIMAL(18, 6)##hash                                 353.29us     2.83K
hash_DECIMAL(18, 6)##xxhash64                             346.79us     2.88K
----------------------------------------------------------------------------
hash_DECIMAL(38, 6)##hash                                 787.34us     1.27K
hash_DECIMAL(38, 6)##xxhash64                             742.86us     1.35K
----------------------------------------------------------------------------
hash_DOUBLE##hash                                         370.08us     2.70K
hash_DOUBLE##xxhash64                                     371.87us     2.69K
----------------------------------------------------------------------------
hash_REAL##hash                                           349.01us     2.87K
hash_REAL##xxhash64                                       360.70us     2.77K
----------------------------------------------------------------------------
hash_ROW<f_map:MAP<INTEGER,VARCHAR>,f_array:ARR            17.70ms     56.50
hash_ROW<f_map:MAP<INTEGER,VARCHAR>,f_array:ARR            17.93ms     55.77
----------------------------------------------------------------------------
hash_SMALLINT##hash                                       331.26us     3.02K
hash_SMALLINT##xxhash64                                   340.99us     2.93K
----------------------------------------------------------------------------
hash_TIMESTAMP##hash                                      493.88us     2.02K
hash_TIMESTAMP##xxhash64                                  486.54us     2.06K
----------------------------------------------------------------------------
hash_TINYINT##hash                                        328.91us     3.04K
hash_TINYINT##xxhash64                                    348.23us     2.87K
----------------------------------------------------------------------------
hash_VARCHAR##hash                                          1.09ms    916.55
hash_VARCHAR##xxhash64                                    959.35us     1.04K
----------------------------------------------------------------------------
```
Benchmark result before after optimization:
```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
hash_ARRAY<MAP<INTEGER,VARCHAR>>##hash                     10.27ms     97.38
hash_ARRAY<MAP<INTEGER,VARCHAR>>##xxhash64                  9.41ms    106.22
----------------------------------------------------------------------------
hash_BIGINT##hash                                         161.41us     6.20K
hash_BIGINT##xxhash64                                     151.50us     6.60K
----------------------------------------------------------------------------
hash_DECIMAL(18, 6)##hash                                 166.50us     6.01K
hash_DECIMAL(18, 6)##xxhash64                             158.21us     6.32K
----------------------------------------------------------------------------
hash_DECIMAL(38, 6)##hash                                 603.90us     1.66K
hash_DECIMAL(38, 6)##xxhash64                             589.37us     1.70K
----------------------------------------------------------------------------
hash_DOUBLE##hash                                         183.81us     5.44K
hash_DOUBLE##xxhash64                                     162.11us     6.17K
----------------------------------------------------------------------------
hash_REAL##hash                                            74.44us    13.43K
hash_REAL##xxhash64                                       130.64us     7.65K
----------------------------------------------------------------------------
hash_ROW<f_map:MAP<INTEGER,VARCHAR>,f_array:ARR            18.18ms     55.00
hash_ROW<f_map:MAP<INTEGER,VARCHAR>,f_array:ARR            17.96ms     55.67
----------------------------------------------------------------------------
hash_SMALLINT##hash                                        70.86us    14.11K
hash_SMALLINT##xxhash64                                   123.17us     8.12K
----------------------------------------------------------------------------
hash_TIMESTAMP##hash                                      270.07us     3.70K
hash_TIMESTAMP##xxhash64                                  257.59us     3.88K
----------------------------------------------------------------------------
hash_TINYINT##hash                                         72.28us    13.84K
hash_TINYINT##xxhash64                                    127.60us     7.84K
----------------------------------------------------------------------------
hash_VARCHAR##hash                                        818.48us     1.22K
hash_VARCHAR##xxhash64                                    660.15us     1.51K
----------------------------------------------------------------------------
```